Grants are whole credits: fix the raise that made one 104.2, and the door that then refused it
The operator could not increase their coordinator's grant — "it just fails
outright", no dialog, nothing in the event log. Reproduced: hiring a
104-credit report on `gpt-reserve` (seat 0.2) under a 100-credit superior
left the superior on exactly 104.2, and off a fractional grant every delta
the credit bar can compute is itself fractional (it rounds its TARGET to a
whole number and sends `target - grant`), while `Op.delta` was typed `int`.
Pydantic answered 422 in BOTH directions.

The ledger's arithmetic was never the problem — it quantises every write and
handled a whole delta fine. What was wrong was the belief, written into a
comment on `switch_model`, that a melt is "THE ONE PATH THAT MAKES A GRANT
FRACTIONAL"; `_chain_acquire` makes them too, and the UI and the ops door
were both built on that false claim.

Per the user's ruling ("just round up grants to the next whole number when
saturating superiors like that; fractional grant amounts is an invalid state
anyway imo"):

  · `_chain_acquire` now snaps each inflated grant UP to a whole credit and
    carries the difference to the node above, so no free() can go negative
    and no parent pays twice. User-actor cascades only — an agent's cascade
    has no pool behind the last carry, and refusing there would break hires
    that work today.
  · a ONE-SHOT load repair (`whole_grants_v1`) rounds existing fractional
    grants UP, deepest-first, so the live 104.2 becomes 105 on first load.
    Once per document on purpose: a melt still lands a seat difference in a
    grant so the node's holding does not move, and a standing round-up rule
    would mint a credit out of the parent on every switch-and-reload cycle.
  · `reallocate` snaps its TARGET up, so every check below it (free credits,
    §4.6 acquire, the D-014 cap) runs on the amount actually written.
  · `Op.delta` is a float — the ledger was always fraction-correct, the int
    was the only thing standing in the way.
  · the credit bar clamps AFTER rounding, against `ceil(committed)`: with a
    fractional floor, a small UPWARD drag used to produce a REDUCTION.

SEATS ARE UNTOUCHED. gpt-reserve and luna still cost 0.2, and `free` is
still fractional whenever such a seat is live — pinned in §4.

test_fractional_grant.py: 27 checks. RED on main in 15 of them, including
the 422 itself; green here. §5 is the negative control — six refusals that
pass on both sides and were each shown red under a mutation that removes the
guard, so they are not decoration.

+++ backend/orgtree/api.py
@@ -6755,7 +6755,19 @@ class Op(Body):
     above: str | None = None
     org_visibility: str | None = None
     effort: str | None = None     # hire — thinking effort, applied WITH the hire
-    delta: int | None = None      # reallocate
+    # reallocate. ⚠ FLOAT, not int, and the difference is a user-visible
+    # outage: the credit bar rounds its TARGET to a whole number and sends
+    # `target - grant`, so against a grant that is fractional for ANY reason
+    # (a switch_model melt still makes one, by design — the node's holding
+    # must not move) every delta it can compute is fractional and pydantic
+    # refused the whole request with `int_from_float`. Measured 2026-09-04 on
+    # the operator's 104.2 coordinator: 422 in BOTH directions, no dialog,
+    # nothing in the event log — "it just fails outright". `reallocate`
+    # itself has always been fraction-correct (it quantises every write), so
+    # the int was the only thing standing in the way. This widens what the
+    # door ACCEPTS, not what the ledger ALLOWS: the free-credit check, the
+    # §4.6 chain acquire and the D-014 top-level cap all still run.
+    delta: float | None = None
     new_parent: str | None = None  # promote / demote
     dir: str | None = None        # revoke_dir
     # ceiling spec §1: the one-action bridge — re-send the same op with this
